Refactoring a few orchestration methods so they can be shared with non-IAM technologies like S3 or even GCP
